Add ability to "record" and "replay" rsc invokations for testing.
The recorder utility in the corresponding package produces a JSON file that contains all the information required to compare two invocations of rsc (process stdout, exit status etc.). This file also contains the information needed to mock the API used to produce it.
The JSON file is consumed by "recording_test.go" to stand up a ghttp server (gomega test server) that emulates the backend API for fast replay.