Fix potential ambiguity when resolving hrefs
If the same href is supported by two actions from different resources
rsc would pick a random one. The algorithm now uses a heuristic where
it favors the action with a matching href that has the same name. It
cannot only consider those though because we do want "fuzzy" matching
in general.